move version back
